# app/api/projects.py
"""
Projects API endpoints for AI Movie Task Service
Follows constitutional requirements for project isolation
"""
from fastapi import APIRouter, Depends, Query, Path
import structlog

from ..models.project import ProjectTasksRequest, ProjectTasksResponse, ProjectTasksSummary
from ..middleware.auth import verify_api_key, validate_project_id

logger = structlog.get_logger()
router = APIRouter()


@router.get("/projects/{project_id}/tasks", response_model=ProjectTasksResponse)
async def get_project_tasks(
    project_id: str = Path(..., description="Project identifier"),
    status: str = Query(None, description="Filter by task status"),
    task_type: str = Query(None, description="Filter by task type"),
    page: int = Query(1, ge=1, description="Page number"),
    limit: int = Query(20, ge=1, le=100, description="Items per page"),
    api_key: str = Depends(verify_api_key)
) -> ProjectTasksResponse:
    """
    Get all tasks for a specific project with filtering and pagination
    Constitutional requirement: Project isolation enforcement
    """
    # Validate project ID format
    validated_project_id = validate_project_id(project_id)
    
    logger.info(
        "Project tasks listing requested",
        project_id=validated_project_id,
        status_filter=status,
        task_type_filter=task_type,
        page=page,
        limit=limit
    )
    
    # Create request model for validation
    request = ProjectTasksRequest(
        status=status,
        task_type=task_type,
        page=page,
        limit=limit
    )
    
    # TODO: Retrieve tasks from Redis with project isolation
    # TODO: Apply filters and pagination
    
    # Return empty result for now
    return ProjectTasksResponse(
        tasks=[],
        pagination={
            "page": page,
            "limit": limit,
            "total": 0,
            "total_pages": 0
        }
    )