# app/api/tasks.py
"""
Tasks API endpoints for AI Movie Task Service
Follows constitutional requirements for API design and security
"""
from datetime import datetime
from uuid import uuid4
from fastapi import APIRouter, Depends, HTTPException, Path
from typing import Dict, Any
import structlog

from ..models.task import (
    TaskSubmissionRequest, 
    TaskSubmissionResponse, 
    TaskStatusResponse,
    TaskStatus,
    Task
)
from ..middleware.auth import verify_api_key, validate_task_id

logger = structlog.get_logger()
router = APIRouter()


@router.post("/tasks/submit", response_model=TaskSubmissionResponse, status_code=201)
async def submit_task(
    task_request: TaskSubmissionRequest,
    api_key: str = Depends(verify_api_key)
) -> TaskSubmissionResponse:
    """
    Submit a new task for processing
    Constitutional requirement: Service-first architecture with clear API boundaries
    """
    logger.info(
        "Task submission received",
        project_id=task_request.project_id,
        task_type=task_request.task_type,
        priority=task_request.priority
    )
    
    # Generate unique task ID
    task_id = uuid4()
    
    # Create task instance
    task = Task(
        task_id=task_id,
        project_id=task_request.project_id,
        task_type=task_request.task_type,
        status=TaskStatus.QUEUED,
        priority=task_request.priority,
        task_data=task_request.task_data,
        callback_url=task_request.callback_url,
        metadata=task_request.metadata,
        created_at=datetime.utcnow(),
        estimated_duration=300,  # Default estimate
        queue_position=1  # TODO: Calculate actual queue position
    )
    
    # TODO: Store task in Redis
    # TODO: Submit to Celery queue
    
    logger.info(
        "Task submitted successfully",
        task_id=str(task_id),
        project_id=task_request.project_id
    )
    
    return TaskSubmissionResponse(
        task_id=task_id,
        status=TaskStatus.QUEUED,
        project_id=task_request.project_id,
        estimated_duration=task.estimated_duration,
        queue_position=task.queue_position,
        created_at=task.created_at
    )


@router.get("/tasks/{task_id}/status", response_model=TaskStatusResponse)
async def get_task_status(
    task_id: str = Path(..., description="Task UUID"),
    api_key: str = Depends(verify_api_key)
) -> TaskStatusResponse:
    """
    Get task status and progress
    Constitutional requirement: Real-time monitoring capabilities
    """
    # Validate task ID format
    validated_task_id = validate_task_id(task_id)
    
    logger.info(
        "Task status requested",
        task_id=validated_task_id
    )
    
    # TODO: Retrieve task from Redis
    # For now, return mock data to satisfy contract
    
    # Mock task not found
    raise HTTPException(
        status_code=404,
        detail="Task not found"
    )


@router.delete("/tasks/{task_id}/cancel")
async def cancel_task(
    task_id: str = Path(..., description="Task UUID"),
    api_key: str = Depends(verify_api_key)
) -> Dict[str, Any]:
    """
    Cancel a queued or processing task
    Constitutional requirement: Task lifecycle management
    """
    # Validate task ID format
    validated_task_id = validate_task_id(task_id)
    
    logger.info(
        "Task cancellation requested",
        task_id=validated_task_id
    )
    
    # TODO: Retrieve task from Redis
    # TODO: Cancel task in Celery
    
    # Mock task not found for now
    raise HTTPException(
        status_code=404,
        detail="Task not found"
    )


@router.post("/tasks/{task_id}/retry", response_model=TaskSubmissionResponse, status_code=201)
async def retry_task(
    task_id: str = Path(..., description="Task UUID"),
    api_key: str = Depends(verify_api_key)
) -> TaskSubmissionResponse:
    """
    Retry a failed task with same parameters
    Constitutional requirement: Reliability with retry mechanisms
    """
    # Validate task ID format
    validated_task_id = validate_task_id(task_id)
    
    logger.info(
        "Task retry requested",
        task_id=validated_task_id
    )
    
    # TODO: Retrieve original task from Redis
    # TODO: Validate task is in failed state
    # TODO: Create new task with same parameters
    
    # Mock task not found for now
    raise HTTPException(
        status_code=404,
        detail="Task not found"
    )